requirements.txt
# Core Dependencies
numpy==1.24.3
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
Pillow==10.0.0

# Deep Learning
tensorflow==2.15.0
tensorflow-hub==0.15.0
keras==2.15.0

# Transfer Learning & Vision
tensorflow-addons==0.22.0

# Head Pose Estimation
mediapipe==0.10.14

# Data Processing
scikit-image==0.21.0
scipy==1.11.1
pandas==2.0.3

# Visualization & Plotting
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
tqdm==4.65.0
pyyaml==6.0
python-dotenv==1.0.0

# Jupyter
jupyter==1.0.0
ipython==8.14.0
notebook==7.0.0

# Testing & Evaluation
scikit-learn==1.3.0

# Logging & Monitoring
loguru==0.7.0

# Kaggle
kagglehub==0.2.0
